+++ backend/src/accounts/accounts.module.ts
@@ -2,10 +2,11 @@ import { HttpModule, Module } from '@nestjs/common';
 
 import { AccountsService } from './accounts.service';
 import { ApiService } from './api.service';
+import { MetricsService } from './metrics.service';
 
 @Module({
   imports: [HttpModule],
-  providers: [AccountsService, ApiService],
+  providers: [AccountsService, ApiService, MetricsService],
   exports: [AccountsService, ApiService],
 })
 export class AccountsModule {}
