--- backend/src/common/interfaces.ts
@@ -1,4 +1,0 @@
-export interface DateInterval {
-  start: Date;
-  end: Date;
-}
